Move spigot package to spigot.latest
