+++ src/main/java/dev/mikchan/mcnp/chat/implementation/spigot/latest/commands/command/ICommand.kt
@@ -1,4 +1,4 @@
-package dev.mikchan.mcnp.chat.implementation.spigot.commands.command
+package dev.mikchan.mcnp.chat.implementation.spigot.latest.commands.command
 
 import org.bukkit.command.CommandExecutor
 import org.bukkit.command.TabCompleter
